+++ gomill_tests/sgf_grammar_tests.py
@@ -105,7 +105,6 @@ def test_tokeniser(tc):
 
     tc.assertEqual(check_incomplete("(;B[ag"), (3, 3))
     tc.assertEqual(check_incomplete("(;B[ag)"), (3, 3))
-    tc.assertEqual(check_incomplete("(;AddBlack[ag])"), (3, 3))
     tc.assertEqual(check_incomplete("(;+B[ag])"), (2, 2))
     tc.assertEqual(check_incomplete("(;B+[ag])"), (3, 3))
     tc.assertEqual(check_incomplete("(;B[ag]+)"), (4, 7))
@@ -116,6 +115,16 @@ def test_tokeniser(tc):
     tc.assertEqual(check_complete(r"(;[ab \] \\\] cd][ef]"), 4)
     tc.assertEqual(check_incomplete(r"(;B[ag\])"), (3, 3))
     tc.assertEqual(check_incomplete(r"(;B[ag\\\])"), (3, 3))
+
+def test_tokeniser_lower_case_propidents(tc):
+    tokenise = sgf_grammar.tokenise
+
+    tc.assertEqual(tokenise(b"(;AddBlack[ag])")[0],
+                   [('D', b'('),
+                    ('D', b';'),
+                    ('I', b'AB'),
+                    ('V', b'ag'),
+                    ('D', b')')])
 
 def test_parser_structure(tc):
     parse_sgf_game = sgf_grammar.parse_sgf_game